Add validation and error handling

+++ src/main/java/dev/mehdi/aftas/dto/fish/FishResponseDto.java
@@ -12,7 +12,7 @@ import lombok.Setter;
 public class FishResponseDto {
     private Long id;
     private String name;
-    private Long averageWeight;
+    private Float averageWeight;
 
     public static FishResponseDto fromModel(Fish fish) {
         return new FishResponseDto(
